Rename RecoEdge --> EnvisEdge (#157)
* rename to --> envisedge

* change logo

* Update README.md

* removed old assets

* Update sparkline-banner.png
