+++ CONTRIBUTING.md
@@ -15,7 +15,7 @@ The library is in the initial stages and we would like to build a complete suite
 
 ## Submitting Pull Requests
 
-Please send a [GitHub Pull Request to RecoEdge](https://github.com/NimbleEdge/RecoEdge) with a clear list of what you've done (read more about [pull requests](http://help.github.com/pull-requests/)). 
+Please send a [GitHub Pull Request to EnvisEdge](https://github.com/NimbleEdge/EnvisEdge) with a clear list of what you've done (read more about [pull requests](http://help.github.com/pull-requests/)). 
 Please Squash your commits into one before sending the pull request. 
 
 Always write a clear log message for your commits. One-line messages are fine for small changes, but bigger changes should look like this:
