+++ scala_core/src/main/scala/org/nimbleedge/envisedge/Main.scala
@@ -1,4 +1,4 @@
-package org.nimbleedge.recoedge
+package org.nimbleedge.envisedge
 
 import models._
