+++ scala_core/src/main/scala/org/nimbleedge/envisedge/RealTimeGraphQuery.scala
@@ -1,6 +1,6 @@
-package org.nimbleedge.recoedge
+package org.nimbleedge.envisedge
 
-import org.nimbleedge.recoedge.models._
+import org.nimbleedge.envisedge.models._
 import scala.concurrent.duration.FiniteDuration
 import akka.actor.typed.Behavior
 import akka.actor.typed.ActorRef
