+++ scala_core/src/main/scala/org/nimbleedge/envisedge/SimulatorSupervisor.scala
@@ -1,4 +1,4 @@
-package org.nimbleedge.recoedge
+package org.nimbleedge.envisedge
 
 import akka.actor.typed.Behavior
 import akka.actor.typed.Signal
